pyproject.toml
[project]
name = "agentsai"
version = "0.1.0"
description = "Agents AI to build content."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.3",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "apify-client>=1.0.0",
    "anyio>=4.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.10",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
